requirements.txt
streamlit==1.20.0
PyMuPDF==1.22.3
img2pdf==0.4.4
Pillow==9.4.0
